+++ src/main/model/Run.java
@@ -9,7 +9,7 @@ public class Run extends Workout {
     // REQUIRES: runName is not an empty string, todaysDate is not an empty string, runHR > 0
     // EFFECTS: creates new run object with given parameters
     public Run(String runName, String todaysDate, int runHR, int runPaceMins, int runPaceSecs, int runTime,
-               int runPerceivedDifficulty, int runDistance) {
+               int runPerceivedDifficulty, double runDistance) {
 
         super(runName, todaysDate, runHR, runTime, runPerceivedDifficulty, runDistance);
         this.avgRunPaceMins = runPaceMins;
